+++ includes/class-tmsm-aquatonic-course-booking.php
@@ -135,6 +135,11 @@ class Tmsm_Aquatonic_Course_Booking {
 		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/class-tmsm-aquatonic-course-booking-sanitize.php';
 
 		/**
+		 * The class responsible for list table
+		 */
+		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/class-tmsm-aquatonic-course-booking-list-table.php';
+
+		/**
 		 * The classes for Dialog Insight (booking & contact)
 		 */
 		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/class-dialog-insight-api.php';