+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Remove debug
 * Updated translations
 * Fix realtime value when negative
+* Better view of bookings, with filters: name, course date, creation date
 
 ### 1.4.5: May 31th, 2021
 * Automatic refresh of data now executed one minute after cron event has executed, to let time to do the actions